Restore head - Add hook for torch.nn.functional, factorize args hooking in a single file for method and function hooking, add unary method tests
