+++ test/torch/test_hook.py
@@ -65,28 +65,63 @@ class TestHook(object):
             assert isinstance(err_pointer, PointerTensor)
             assert err_pointer.id == ptr_id
 
-    @pytest.mark.parametrize("attr", ["add", "mul"])
-    def test_get_pointer_method(self, attr):
+    @pytest.mark.parametrize("attr", ["abs"])
+    def test_get_pointer_unary_method(self, attr):
         self.setUp()
         x = torch.Tensor([1, 2, 3])
-        method = getattr(x, attr)
         native_method = getattr(x, f"native_{attr}")
-        expected = native_method(x)
+        expected = native_method()
         x_ptr = x.send(self.bob)
-        res_ptr = method(x_ptr)
+        method = getattr(x_ptr, attr)
+        res_ptr = method()
         res = res_ptr.get()
         assert (res == expected).all()
 
     @pytest.mark.parametrize("attr", ["add", "mul"])
-    def test_get_pointer_to_pointer_method(self, attr):
+    def test_get_pointer_binary_method(self, attr):
         self.setUp()
         x = torch.Tensor([1, 2, 3])
-        method = getattr(x, attr)
+        native_method = getattr(x, f"native_{attr}")
+        expected = native_method(x)
+        x_ptr = x.send(self.bob)
+        method = getattr(x_ptr, attr)
+        res_ptr = method(x_ptr)
+        res = res_ptr.get()
+        assert (res == expected).all()
+
+    @pytest.mark.parametrize("attr", ["abs"])
+    def test_get_pointer_to_pointer_unary_method(self, attr):
+        self.setUp()
+        x = torch.Tensor([1, 2, 3])
+        native_method = getattr(x, f"native_{attr}")
+        expected = native_method()
+        x_ptr = x.send(self.bob).send(self.alice)
+        method = getattr(x_ptr, attr)
+        res_ptr = method()
+        res = res_ptr.get().get()
+        assert (res == expected).all()
+
+    @pytest.mark.parametrize("attr", ["add", "mul"])
+    def test_get_pointer_to_pointer_binary_method(self, attr):
+        self.setUp()
+        x = torch.Tensor([1, 2, 3])
         native_method = getattr(x, f"native_{attr}")
         expected = native_method(x)
         x_ptr = x.send(self.bob).send(self.alice)
+        method = getattr(x_ptr, attr)
         res_ptr = method(x_ptr)
         res = res_ptr.get().get()
+        assert (res == expected).all()
+
+    @pytest.mark.parametrize("attr", ["relu", "celu", "elu"])
+    def test_hook_module_functional(self, attr):
+        self.setUp()
+        attr = getattr(F, attr)
+        x = torch.Tensor([1, -1, 3, 4])
+        expected = attr(x)
+        x_ptr = x.send(self.bob)
+        res_ptr = attr(x_ptr)
+        res = res_ptr.get()
         assert (res == expected).all()
 
     def test_properties(self):
